refactor(yomu): query.rs から keyword/rank モジュールを抽出 (#137)
RC-009 (#137) の file 分割。query.rs 558→283 行に縮小し、#255 の
file-size gate allowlist から query.rs を除去する。

- keyword 抽出 (stop-word/stem/type-hint) → src/query/keywords.rs (97 行)
- ranking (rerank/RerankContext/keyword_hit_ratio/cap_per_file/
  cross_encoder/is_test_path/semantic_confidence) → src/query/rank.rs (195 行)
- 型 (QueryError/SearchStages/SearchOutcome) + orchestration
  (search/search_pipeline/search_from_file/capture_stage) は query.rs に残す

公開 API (extract_keywords / extract_type_hints / rerank / RerankContext /
search / search_from_file / SearchOutcome / QueryError) は pub use 再エクスポート
で呼び出し側 (tools/*) 不変。内部 helper は pub(super) とし (LANG.md 最小可視性)、
query/tests.rs は super::rank::{} で明示 import (既存 tools/tests.rs・indexer/tests.rs 慣習)。

挙動保存: 718 tests pass (4 skipped、ベースライン同一)。diff-cover 97%
(rank.rs の早期 return ガード 2 行と cross_encoder_rerank のエラーアーム
2 行は分割前から未カバー、>=95% floor は満たす)。

Refs #137
